requirements.txt
# LangChain Ecosystem
langchain==0.1.3
langchain_openai==0.0.5
langchain-community
langchain-core

# Vector Databases & Similarity Search
chromadb==0.5.0
faiss-cpu

# OpenAI
openai==1.10
tiktoken>=0.5.2

# UI Framework
gradio==4.44.1
setuptools>=65.5.1

# Data Processing & Utils
numpy==1.26.2
pandas==2.1.3
python-dotenv==1.0.0

# Core Dependencies
#pydantic==2.5.2
typing-extensions==4.8.0
requests==2.31.0